update from upstream

+++ src/ackermann_function.rs
@@ -1,6 +1,6 @@
-// Implements http://rosettacode.org/wiki/Ackermann_function
+// http://rosettacode.org/wiki/Ackermann_function
 
-fn ack(m: int, n: int) -> int {
+fn ack(m: isize, n: isize) -> isize {
     if m == 0 {
         n + 1
     } else if n == 0 {
@@ -17,7 +17,6 @@ fn test_ack() {
     assert_eq!(expected, found)
 }
 
-#[cfg(not(test))]
 fn main() {
     let a = ack(3, 4);
     println!("{}", a);
