+++ src/sum_digits.rs
@@ -1,6 +1,7 @@
-// Implements http://rosettacode.org/wiki/Sum_digits_of_an_integer
+// http://rosettacode.org/wiki/Sum_digits_of_an_integer
 
-fn sum(n: uint, base: uint) -> uint {
+
+fn sum(n: usize, base: usize) -> usize {
     let mut total = 0;
     let mut n = n;
     while n != 0 {
@@ -22,7 +23,6 @@ fn base_16() {
     assert_eq!(sum(0xf0e, 16), 29);
 }
 
-#[cfg(not(test))]
 fn main() {
     println!("{}", sum(1234, 10));
 }
